integrated ble

+++ main/core/include/constants.hpp
@@ -3,5 +3,5 @@
 enum ActionStatus { OK, INVALID_TIME };
 
 using Pin = uint8_t;
-using Esp_Err_t = uint8_t;
+using Esp_Err_t = int;
 using byte_count = int;
